rest plugin decoupling
