+++ services/web/server/src/simcore_service_webserver/catalog_api_handlers.py
@@ -22,7 +22,7 @@ from .catalog_api_models import (
 )
 from .constants import RQ_PRODUCT_KEY
 from .login.decorators import RQT_USERID_KEY, login_required
-from .rest_utils import RESPONSE_MODEL_POLICY
+from .rest_constants import RESPONSE_MODEL_POLICY
 from .security_decorators import permission_required
 
 ###############
