+++ services/web/server/src/simcore_service_webserver/projects/projects_handlers.py
@@ -23,7 +23,7 @@ from .._meta import api_version_prefix as VTAG
 from ..constants import RQ_PRODUCT_KEY
 from ..login.decorators import RQT_USERID_KEY, login_required
 from ..resource_manager.websocket_manager import PROJECT_ID_KEY, managed_resource
-from ..rest_utils import RESPONSE_MODEL_POLICY
+from ..rest_constants import RESPONSE_MODEL_POLICY
 from ..security_api import check_permission
 from ..security_decorators import permission_required
 from ..storage_api import copy_data_folders_from_project
