+++ services/web/server/src/simcore_service_webserver/rest.py
@@ -6,49 +6,25 @@
 
 """
 import logging
-from functools import lru_cache
-from pathlib import Path
-from typing import Optional
+from typing import Tuple
 
-import openapi_core
-import yaml
 from aiohttp import web
 from aiohttp_swagger import setup_swagger
-from openapi_core.schema.specs.models import Spec as OpenApiSpecs
 from servicelib.aiohttp import openapi
 from servicelib.aiohttp.application_setup import ModuleCategory, app_module_setup
 from servicelib.aiohttp.rest_middlewares import (
     envelope_middleware_factory,
     error_middleware_factory,
 )
-from simcore_service_webserver.resources import resources
 
 from . import rest_routes
 from ._meta import api_version_prefix
 from .constants import APP_OPENAPI_SPECS_KEY
 from .diagnostics_config import get_diagnostics_config
 from .rest_config import assert_valid_config
+from .rest_utils import get_openapi_specs_path, load_openapi_specs
 
 log = logging.getLogger(__name__)
-
-
-def get_openapi_specs_path(api_version_dir: Optional[str] = None) -> Path:
-    if api_version_dir is None:
-        api_version_dir = api_version_prefix
-
-    return resources.get_path(f"api/{api_version_dir}/openapi.yaml")
-
-
-@lru_cache  # required to boost tests speed, gains 3.5s per test
-def load_openapi_specs(spec_path: Optional[Path] = None) -> OpenApiSpecs:
-    if spec_path is None:
-        spec_path = get_openapi_specs_path()
-
-    with spec_path.open() as fh:
-        spec_dict = yaml.safe_load(fh)
-    specs: OpenApiSpecs = openapi_core.create_spec(spec_dict, spec_path.as_uri())
-
-    return specs
 
 
 @app_module_setup(
@@ -57,7 +33,7 @@ def load_openapi_specs(spec_path: Optional[Path] = None) -> OpenApiSpecs:
     depends=["simcore_service_webserver.security"],
     logger=log,
 )
-def setup(app: web.Application, *, swagger_doc_enabled: bool = True):
+def setup_rest(app: web.Application, *, swagger_doc_enabled: bool = True):
     # ----------------------------------------------
     # TODO: temporary, just to check compatibility between
     # trafaret and pydantic schemas
@@ -115,7 +91,4 @@ def setup(app: web.Application, *, swagger_doc_enabled: bool = True):
         )
 
 
-# alias
-setup_rest = setup
-
-__all__ = ["setup_rest"]
+__all__: Tuple[str, ...] = ("setup_rest",)
