+++ services/web/server/src/simcore_service_webserver/rest_constants.py
@@ -1,0 +1,8 @@
+# SEE https://pydantic-docs.helpmanual.io/usage/exporting_models/#modeldict
+
+RESPONSE_MODEL_POLICY = {
+    "by_alias": True,
+    "exclude_unset": True,
+    "exclude_defaults": False,
+    "exclude_none": False,
+}
