+++ services/web/server/src/simcore_service_webserver/rest_utils.py
@@ -1,8 +1,29 @@
-# SEE https://pydantic-docs.helpmanual.io/usage/exporting_models/#modeldict
+from functools import lru_cache
+from pathlib import Path
+from typing import Optional
 
-RESPONSE_MODEL_POLICY = {
-    "by_alias": True,
-    "exclude_unset": True,
-    "exclude_defaults": False,
-    "exclude_none": False,
-}
+import openapi_core
+import yaml
+from openapi_core.schema.specs.models import Spec as OpenApiSpecs
+from simcore_service_webserver.resources import resources
+
+from ._meta import api_version_prefix
+
+
+def get_openapi_specs_path(api_version_dir: Optional[str] = None) -> Path:
+    if api_version_dir is None:
+        api_version_dir = api_version_prefix
+
+    return resources.get_path(f"api/{api_version_dir}/openapi.yaml")
+
+
+@lru_cache  # required to boost tests speed, gains 3.5s per test
+def load_openapi_specs(spec_path: Optional[Path] = None) -> OpenApiSpecs:
+    if spec_path is None:
+        spec_path = get_openapi_specs_path()
+
+    with spec_path.open() as fh:
+        spec_dict = yaml.safe_load(fh)
+    specs: OpenApiSpecs = openapi_core.create_spec(spec_dict, spec_path.as_uri())
+
+    return specs
