+++ services/web/server/src/simcore_service_webserver/version_control_handlers.py
@@ -8,7 +8,7 @@ from pydantic.decorator import validate_arguments
 
 from ._meta import api_version_prefix as VTAG
 from .login.decorators import login_required
-from .rest_utils import RESPONSE_MODEL_POLICY
+from .rest_constants import RESPONSE_MODEL_POLICY
 from .security_decorators import permission_required
 from .utils_aiohttp import (
     create_url_for_function,
